Use assertions in code
Add java assert statements at various points of the code to document important assumptions.

In Parser, assert that the date fields for deadline and event tasks should not be null, and check that the parsed idx follows 1-based indexing.

In Storage, assert that directories and filepaths are valid and are created.

 In Task, assert that task descriptions and types are not null.